HD44780: Initial commit
This commit introduces the HD44780 display decoder for displays based on the common HITACHI
and compatible controllers. It uses 4 or 8 data bits and requires additionally at least
a RS and an EN bit, the R/W bit is optional. It supports the generation of png images
representing the current content of the display, which works, but for now produces errors
when used in Pulseview.